haidz2/requirements.txt
# Core dependencies
playwright==1.40.0
pydantic==2.5.3
pandas==2.1.4
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
aiohttp==3.9.1
requests>=2.31.0

# AI/LLM dependencies (optional - for true AI brain)
openai>=1.0.0
tenacity>=8.0.0

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development dependencies
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Type stubs
pandas-stubs==2.1.4.231227
types-requests==2.31.0.20240106